third_party/blink/web_tests/platform/win/external/wpt/offscreen-canvas/text/2d.text.measure.fontBoundingBox-expected.txt

This is a testharness.js-based test.
FAIL Testing fontBoundingBox for OffscreenCanvas assert_equals: ctx.measureText('A').fontBoundingBoxAscent === 85 (got 45[number], expected 85[number]) expected 85 but got 45
Harness: the test ran to completion.
